+++ packages/keycloak/src/keycloak.module.ts
@@ -1,11 +1,12 @@
-import { Module } from '@nestjs/common'
+import { Module, HttpModule } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { KeycloakClient } from './client'
-import { ValidateAccessTokenService } from './service'
+import { GetAccessTokenService, ValidateAccessTokenService } from './service'
 import { KeycloakStrategy } from './strategy'
 
 @Module({
-  imports: [ConfigModule.forRoot({ isGlobal: true })],
-  providers: [KeycloakClient, ValidateAccessTokenService, KeycloakStrategy],
+  imports: [ConfigModule.forRoot({ isGlobal: true }), HttpModule],
+  exports: [GetAccessTokenService],
+  providers: [KeycloakClient, ValidateAccessTokenService, GetAccessTokenService, KeycloakStrategy],
 })
 export class KeycloakModule {}