+++ packages/keycloak/src/service/index.ts
@@ -1,1 +1,2 @@
 export * from './validate-access-token.service'
+export * from './get-access-token.service'